add restricted pages

+++ base/views.py
@@ -2,6 +2,7 @@ from django.shortcuts import render, redirect
 from django.db.models import Q
 from django.http import HttpResponse
 from .models import Room, Topic
+from django.contrib.auth.decorators import login_required
 from .forms import RoomForm
 from django.contrib.auth import authenticate, login, logout
 from django.contrib import messages
@@ -46,6 +47,7 @@ def room(request, id):
     context = {'room' : room}
     return render(request, 'base/room.html', context)
 
+@login_required(login_url='/login')
 def createRoom(request):
     form = RoomForm
     if request.method == 'POST':
@@ -57,9 +59,13 @@ def createRoom(request):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='/login')
 def updateRoom(request, id):
     room = Room.objects.get(id=id)
     form = RoomForm(instance=room)
+    
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here')
     
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -69,9 +75,13 @@ def updateRoom(request, id):
     
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
-    
+
+@login_required(login_url='/login')  
 def deleteRoom(request, id):
     room = Room.objects.get(id=id)
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here')
+    
     if request.method == 'POST':
         room.delete()
         return redirect('home')
